move IssueCheckConfig to a new "environment" app so it's not a tenant schema
