+++ EquiTrack/management/tests/test_issue_checks.py
@@ -2,11 +2,11 @@ from django.core.exceptions import ImproperlyConfigured
 from django.test import override_settings
 from EquiTrack.factories import PartnerFactory
 from EquiTrack.tests.mixins import FastTenantTestCase
+from environment.models import IssueCheckConfig
 from management.issues.checks import BaseIssueCheck, get_issue_checks, get_issue_check_by_id, run_all_checks, \
     ModelCheckData, bootstrap_checks, get_active_issue_checks
 from management.issues.exceptions import IssueFoundException, IssueCheckNotFoundException
-from management.models import FlaggedIssue, ISSUE_STATUS_NEW, ISSUE_STATUS_RESOLVED, ISSUE_STATUS_REACTIVATED, \
-    IssueCheckConfig
+from management.models import FlaggedIssue, ISSUE_STATUS_NEW, ISSUE_STATUS_RESOLVED, ISSUE_STATUS_REACTIVATED
 from partners.models import PartnerOrganization
